requirements.txt
openwakeword==0.6.0
sounddevice==0.4.6
numpy==1.26.4
soundfile>=0.12.1
faster-whisper>=1.0.0
requests>=2.28.0
PyYAML>=6.0
piper-tts>=1.4.0
